refactor: move summary to component
- Move summary cards (SummaryCards) to a separate component

- Create SummaryCards component to organize Incoming, Outgoing and Current Balance cards
